Implement first attempt of AST flattening
The idea, hopefully, is to allow the AST nodes to be packed in a tighter format. Though this is slightly more work, it makes more sense in terms of the amount of space that we're wasting with each node. Just switching to a more constrained array-based storage allows us to save 8 bytes each in 64-bit mode, a 25% increase in density. We'll see how this goes.