Remove Path node type, fix function parsing (partially)
Function parsing currently suffers from the fact that the grammar currently does not distinguish between a function type statement and a function literal. This was originally thought to be a boon, but it now appears that it makes parsing function type def's impossible without some sort of post-parse 'fixing' pass, which is very undesireable. No solution has yet made itself known.